add combined classifier for findsign (contour correlation + entropy).
entropy is pretty sucky though, maybe just use it as a binary filter?
i should come up with better sub-classifiers.

i think i'm also going to start needing a contour manager or things
will get inefficient.  e.g., a good classifier for signs is whether
they contain letters.  but then we probably don't want to try to
classify every contour in the image as a particular letter up front,
but maybe just have a loose classifier we run on all the contours in
the edge-detected image up front to say whether it looks like it could
be a letter, then we can use those later for classifying signs.