get ML module sort of working (30%).
It has been difficult to get good results with linear classification,
so I tried logistic regression instead.  With regularization the
results are 30% correct classification, which isn't super, but
hopefully I can tweak it to do better.  It also classifies noise under
the minimum for any character classification, so its use as a filter
is promising.